udpated deploy script to not generate APP_KEY on successive runs (#14)

+++ deploy-joyenergy.php
@@ -20,19 +20,36 @@ function is_executable_available(string $filename): bool
     }
     return false;
 }
-$res = 0;
+function isEnvValueSet($envKey, $fname): bool
+{
+    $str = file_get_contents($fname);
+    $str .= "\n"; // In case the searched variable is in the last line without \n
+    $keyPosition = strpos($str, "{$envKey}=");
+    $endOfLinePosition = strpos($str, PHP_EOL, $keyPosition);
+    $line = substr($str, $keyPosition, $endOfLinePosition - $keyPosition);
+    $value = substr($line, strpos($line, "=") + 1, strlen($line));
+    if ($value === '') {
+        return False;
+    } else return True;
+}
+
 if (is_executable_available("composer")) {
     echo "Installing Laravel, related packages and setting things up...\n";
     exec("composer install", $output, $res);
-    exec("cp .env.example .env", $output, $res);
-    if ($res !== 0) {
-        echo "Error copying 'env' file...\n";
-        exit;
+    if (!file_exists(".env")) {
+        exec("cp .env.example .env", $output, $res);
+        if ($res !== 0) {
+            echo "Error copying 'env' file...\n";
+            exit;
+        }
     }
-    exec("php artisan key:generate", $output1, $res1);
-    if ($res1 !== 0) {
-        echo "Error generating app key...\n";
-        exit;
+    if (!isEnvValueSet("APP_KEY", ".env")) {
+        echo "Generating the APP KEY...\n";
+        exec("php artisan key:generate", $output1, $res1);
+        if ($res1 !== 0) {
+            echo "Error generating app key...\n";
+            exit;
+        }
     }
     system("php artisan serve");
 } else {
